+++ HISTORY.rst
@@ -4,6 +4,8 @@ History
 
 X.Y.Z (YYYY-MM-DD)
 ------------------
+* Add test case demonstrating the feasibility of parallel writes to
+  Tiled Storage Manager Columns (:pr:`69`)
 * Update README.rst to show that it is possible to convert
   unconstrained columns (ndim==-1) whose rows have the same rank (:pr:`74`)
 * Guard against python-casacore imports (:pr:`73`)